Move react-jsx-development implementation into react-jsx
